backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
psycopg2-binary==2.9.9
supabase==2.0.2
requests==2.31.0
openai>=1.99.0

# ML dependencies (optional - install separately if needed)
# sentence-transformers>=2.2.0
# torch>=2.0.0
# transformers>=4.30.0
# numpy>=1.21.0
# pandas>=1.3.0
